inze/requirements.txt
django~=3.1.5
djangorestframework~=3.12.2
django-cors-headers~=3.6.0
pytest-django~=4.1.0
pytest-mock~=3.5.1
factory-boy~=3.2.0